update inception embedding

+++ tests/test_embeddings.py
@@ -29,14 +29,14 @@ def test_pca():
 
 def test_inception():
     dim = (10, 3, 10, 10)
-    arr = np.random.rand(*dim)
+    arr = np.random.rand(*dim).astype(np.float32)
     embed = embeddings.inception(arr, batch_size=5, device="cpu")
     assert embed.shape == (arr.shape[0], 2048)  # tests correct dims
 
 
 def test_inception_pca():
     dim = (10, 3, 10, 10)
-    arr = np.random.rand(*dim)
+    arr = np.random.rand(*dim).astype(np.float32)
     num_components = 2
     embed = embeddings.inception_pca(
         arr, batch_size=5, device="cpu", n_components=num_components
@@ -46,7 +46,7 @@ def test_inception_pca():
 
 def test_inception_umap():
     dim = (10, 3, 10, 10)
-    arr = np.random.rand(*dim)
+    arr = np.random.rand(*dim).astype(np.float32)
     num_components = 2
     embed = embeddings.inception_umap(
         arr, batch_size=5, device="cpu", random_seed=42, n_components=num_components
